Make dict button switch configurations at boot
Since it's very hard to change the configuration size without changing a
lot of the QMK/TMK code, the easiest way is to only enable either the
MSC or normal keyboard device at boot time. Holding the button down when
booting now enumerates the device as only having a MSC device, and only
sends the corresponding configuration, and also doesn't run the normal
keyboard tasks. Not holding the button down only sends the normal
configuration.